Fix alert policies check (#55)
When deleting the whole alert_policies list from SLO resource, the change is not noticed and therefore, alert policies are not deleted from the resource.
The internal Terraform GetChange function fails to notice a change when a resource is reapplied after unsetting an optional list resource. According to the documentation:

    // HasChange should be used to check if a change exists. It is possible
    // that both the old and new value are the same if the old value was not
    // set and the new value is. This is common, for example, for boolean
    // fields which have a zero value of false.

This commit introduces a fix to the logic of checking the diff on alert policies, that forces to check the difference in size of arrays. Values for this check are gathered from DiffSuppressFunc, and not from GetChange function.