reuse n9 client
Every new N9 client calls Okta for a new auth token. Because of that,
 we're hitting Okta's rate limits.

Now N9 client is created once on the first call and cached for
subsequent calls to the API.

We took into consideration using the client's pool in the provider
and changing SDK to support more concurrent connections, but now
it seems to be a bit over-engineering since TF by default only
runs 10 concurrent operations.

Resolves: PC-6512