chore: Improve docs for target and value fields in SLO resources (#343)
## Motivation

While creating my first multiple-objective ratio SLOs with Terraform, I
was surprised to receive an error on `apply` which required the
`objective.value` field to be unique. I was even more surprised when I
read the module documentation, which informs me that `target` is a
"Designated Value" and `value` is a... "Value".

I see from the YAML Guide that `target` is the objective's percentage
target, and `value` is the threshold value, and that for ratio SLOs
`value` is not used but must be unique across objectives for legacy
reasons. (Computers, right?)

So, here's a doc update that notes all of that in the Terraform
instructions.

## Summary

Updates descriptions for `target` and `value` fields in `nobl9_slo`
`objective`s and `composite`s.

## Related Changes

None.

## Testing

Doc change only.