feat: Switch Service resource to Terraform Plugin Framework [PC-14988] (#351)
## Motivation

The main reason for this transition was the inability to present custom
warnings during Terraform plan.
We needed to present such diagnostic to warn users about Service
recreation and cascade delete of all assigned SLOs which is an
underlying side effect of Service deletion.
The old SDKv2 does not expose any API to do that while the new Plugin
Framework does.

## Summary

How does it all work?

### Multiplexing Providers

There are (and will be until we fully transition to framework) now two,
separate provider definitions. Each one is fully autonomous from each
other.
They are both managed in `main.go` by
[terraform-plugin-mux](https://github.com/hashicorp/terraform-plugin-mux),
which allows multiple providers built with different SDKs to work with
each other.
Terraform mux ensures both providers have the exact same schema.

### Statically typed models

One of the big improvements which the framework brings to the table is
that it no longer operates on generic `map[string]any` when it comes to
schema models.
Instead, users define their own models which can be used to decode the
state into, e.g.:

```go
type ServiceResourceModel struct {
	Name        string            `tfsdk:"name"`
	DisplayName types.String      `tfsdk:"display_name"`
	Project     string            `tfsdk:"project"`
	Description types.String      `tfsdk:"description"`
	Annotations map[string]string `tfsdk:"annotations"`
	Labels      Labels            `tfsdk:"label"`
	Status      types.Object      `tfsdk:"status"`
}
```

### DiffSuppressFunc

The `DiffSuppressFunc` is not available in the plugin framework, instead
developers can define custom types and define custom equality checks on
them.

The `DiffSuppressFunc` was defined for labels and was the main mechanism
which helped us limit the generated diffs in the SDK.
The equality check was not robust enough to me so I decided to define a
custom sorting function for lists `sortListBasedOnReferenceList`, this
function takes a list and a reference list and tries to sort the first
list in the same order that the reference list is sorted, matching the
values based on configurable comparison function. All the elements which
are not matched are appended to the end of the slice.

### Nobl9 SDK Client

Each resource can implement `resource.ResourceWithConfigure` in order to
access whatever thing provider sets using it's own `Configure` function.
This way provider can instantiate Nobl9 `sdk.Client` once and pass it to
all the resources. No need for global state ;)

### Sets vs lists

One of the new types for schema attributes and blocks we can use are
sets, which are essentially unordered, unique-element lists.
This would be ideal for labels IF AND ONLY IF we could define a unique
identifier for each element of the set.
However, that's not possible, instead the framework compares each
element in its entirety. This means more complex objects, like labels
block, must be exactly equal, otherwise, if you change one value (of
`values` attribute) the entire label will be shown in diff as deleted
and an updated version as created.

## Testing

The new framework comes with a [new library for
testing](https://github.com/hashicorp/terraform-plugin-testing).
I've written the new acceptance tests for Service resource using the
aforementioned library.

## Release Notes

Service resource plan was improved, invalid `status` computed attribute
changes will no longer be presented.
Added warning to Service plan when the resource is recreated due to
`name` or `project` which informs the practitioners that the associated
SLOs will be removed and their data lost during recreation.