feat: Add Project Data Source (#409)
## Motivation

In preparation for adding Service as requested by
https://github.com/nobl9/terraform-provider-nobl9/issues/396.
I want to pave the path with Project Data Source but also, it would be
fairly natural to have both Service and Project Data Sources defined,
and not just the one for Service.

I have decided to expose only the Project's `name` for now. We can
always add the rest of the fields If needed.

## Testing

Added dedicated test case for creating a Project in the API and then
using it as Data Source in Terraform to create a Service.

Dispatched tests results:
https://github.com/nobl9/terraform-provider-nobl9/actions/runs/15418917753

## Release Notes

Added Project Data Source. Example:
```tf
data "nobl9_project" "this" {
  name = "my-project"
}
```