feat: Add Service data source (#453)
## Motivation

Solves https://github.com/nobl9/terraform-provider-nobl9/issues/396.
I have decided to expose only the Service's name and project for now. We
can always add the rest of the fields If needed.

## Testing

Added dedicated test case for creating a Service in the API and then
using it as Data Source in Terraform to create a Service in another
Project.

Tests dispatch:
https://github.com/nobl9/terraform-provider-nobl9/actions/runs/15522630236

## Release Notes

Added Service Data Source. Example:

```hcl
data "nobl9_service" "this" {
  name = "my-service"
  project = "my-project"
}
```