fix: Correct version management (#460)
## Motivation

Currently, version reporting does not work, there was a typo in
goreleaser config which had:

```diff
@@ -1 +1 @@
-main.version
+main.Version
```

Due to that, the `Version` variable was always empty and the reported
user agent was always `terraform-`.

## Summary

In order to address this issue and align the approach we have there with
sloctl, I've copied
https://github.com/nobl9/sloctl/blob/main/internal/version.go contents
and adjusted Makefile and goreleaser config.
I've also added `-version` flag which can be used to quickly checkout
the Provider version.

## Related Changes

The code requires adjustment in the Nobl9 API.

## Testing

Release candidate version was built here: . Requests run with it
reported the correct User-Agent.

Tested locally after `make build`:

```bash
└> ./bin/terraform-provider-nobl9 -version
terraform/0.42.0-- (linux amd64 go1.24.4)
```

## Release Notes

Terraform Provider will now report the correct User-Agent. Version of
the Provider may now be verified with `-version` flag.