feat: Add move SLO support [PC-16858] (#466)
## Motivation

In Terraform, we are able to recognize that the SLO's Project is changed
(unlike sloctl), as the resource is identified by its name.
The operation of changing the Project of a given SLO resource can be
then implemented organically, without a need for a dedicated, custom
provider function. When a practitioner changes SLO's `project`
attribute, the SLO will be moved to that new Project, retaining all its
data.

## Summary

Changing the SLO's `project` attribute is now a standard `update`
operation and does not recreate the resource.


![image](https://github.com/user-attachments/assets/25b25e76-b57e-4df5-8949-64559d197ebc)

### Limitations

1. Since under the hood, in the API, moving SLO between Projects is a
dedicated operation, separate from `apply`, we MUST NOT allow the user
to change other attributes. If a practitioner tries to do that, they
will receive the following error:


![image](https://github.com/user-attachments/assets/f8f7a360-d4ce-4331-8287-74b3f745b026)

2. Currently, moving SLOs between Projects is not allowed If an SLO has
any Alert Policies attached. If a practitioner tries to do that, they
will receive the following error:


![image](https://github.com/user-attachments/assets/8a6fe02d-4335-4561-b400-3924a430e418)

## Related Changes

https://github.com/nobl9/nobl9-go/pull/739

## Testing

Added tests which cover validation errors and the project change.
Tests dispatch:
https://github.com/nobl9/terraform-provider-nobl9/actions/runs/16168513749/job/45635995091

### Plan checker

I improved the plan checker utility, it can now be configured to freely
check if the plan contains any added, removed or modified attributes in
the root level.

## Release Notes

Changing SLO's name will no longer result in the resource's recreation,
but instead will move the SLO between the Projects.
Furthermore, changing `project` attribute is only allowed if no other
attributes are changed and when no Alert Policies are attached to the
moved SLO.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>