feat: add Terraform pre-installation step to acceptance tests workflow
This change introduces a pre-installation step for Terraform in the acceptance tests workflow. The addition addresses
a known issue where parallel acceptance tests might fail if Terraform is not pre-installed. The `hashicorp/setup-terraform@v3`
action is used with the `terraform_wrapper` option set to `false`. This ensures compatibility and stability during the
execution of the tests.